security(report): derive the PDF report from the caller's own holdings (#594)
POST /api/v1/report rendered whatever DashboardData the client posted. Any
authenticated user could put invented figures into a report that looks like the
app produced it, and could hand the server an arbitrarily large or deeply nested
object to render — the endpoint had no schema, no size limit and no rate limit,
and rendering a PDF is the most expensive thing this app does per request.

The overview computation moves out of GET /api/v1/dashboard/overview into
lib/dashboardOverview, so the report route can build the same numbers itself
from the caller's own holdings. The two routes are now auth + transport around
one shared builder.

What the client may still send is a locale, allowlisted against lib/locale
(which i18n/request now shares) — anything else is a 400. The body is capped at
256 bytes via a new `maxBytes` option on readJsonBody, which refuses on
Content-Length before reading and re-measures the decoded bytes after; that is
what bounds oversized and deeply nested payloads. A per-user fixed window (10
renders / 60s) in Postgres bounds the rest, mirroring the gold-refresh and NAV
limiters and failing closed if the check is unavailable.

Tests: route tests for forged data, oversized and deeply nested bodies, the
locale allowlist, the limit and its fail-closed path; readJsonBody maxBytes
units; a SQL test for the limiter's policy, per-user isolation and grants.
