+++ lib/locale.ts
@@ -9,3 +9,20 @@
 export function setLocaleCookie(next: string): void {
   document.cookie = `locale=${next};path=/;max-age=31536000;SameSite=Lax`
 }
+
+/**
+ * The two locales the app ships messages for. The default is Vietnamese —
+ * an unrecognised value falls back to it rather than 404-ing a translation.
+ */
+export const APP_LOCALES = ['vi', 'en'] as const
+export type AppLocale = (typeof APP_LOCALES)[number]
+export const DEFAULT_LOCALE: AppLocale = 'vi'
+
+/**
+ * Allowlist check for a locale that arrived from outside — a cookie, a query
+ * string, a request body. The PDF report route takes a locale from the client
+ * and nothing else, so this is the whole of what it will accept (#594).
+ */
+export function isAppLocale(value: unknown): value is AppLocale {
+  return typeof value === 'string' && (APP_LOCALES as readonly string[]).includes(value)
+}
